backend/requirements.txt
# Core LangGraph + LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langchain-mcp-adapters>=0.1.0

# MCP
fastmcp>=2.0.0

# FastAPI backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0
python-dotenv>=1.0.0

# External search/recipe APIs
tavily-python>=0.5.0
httpx>=0.27.0

# Fuzzy matching
rapidfuzz>=3.9.0

# Observability
langsmith>=0.1.0

# Utilities
pydantic>=2.7.0

# Dev / test
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
ruff>=0.4.0
